+++ jsolex/src/main/java/me/champeau/a4j/jsolex/app/listeners/Spectral3DVisualizationHelper.java
@@ -359,7 +359,6 @@ final class Spectral3DVisualizationHelper {
                             BatchOperations.submitOneOfAKind("progress", () -> owner.updateProgress(e.getPayload()));
                         }
                     });
-                    solexVideoProcessor.setIgnoreIncompleteShifts(true);
                     if (mainEllipse != null) {
                         solexVideoProcessor.setCachedEllipse(mainEllipse);
                     }